+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * New interface `build_modes` (`vibrational`) can construct the vibrational energies and normal coordinates from the force constants matrix.
 * Most conversion functions in `PhysFact` (`physics`) support the reverse operation, with the optional keyword `reverse`.
 * New conversion function from atomic unit of mass to unified atomic mass in `PhysFact` (`physics`).
+* New procedure `set_orientation` (`vibrational`) to fix the orientation of normal modes, setting the largest component to positive.
 
 ### Fixed
 * `prt_coord` (`output`) failed if masses were not provided.